# go/docs/swagger.yaml
host: http://localhost:8081/swagger/index.html
info:
  contact:
    email: zhou851708184@163.com
    name: 周凯
    url: http://www.start6.cn
  description: 后台管理
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: vue3-go-admin 后台管理接口文档
  version: "1.0"
paths:
  /api/v1/depart:
    get:
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: page_size
        in: query
        name: page_size
        type: string
      responses:
        "200":
          description: ok
          schema:
            type: string
      tags:
      - 部门
  /api/v1/user:
    delete:
      description: 删除用户
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      tags:
      - 用户
    get:
      description: 获取用户列表
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: page_size
        in: query
        name: size
        type: string
      responses:
        "200":
          description: ok
          schema:
            type: string
      tags:
      - 用户
    post:
      description: 创建用户
      parameters:
      - description: 用户登录账号
        in: body
        name: Account
        required: true
        schema:
          type: string
      - description: 密码
        in: body
        name: PassWord
        required: true
        schema:
          type: string
      - description: 用户显示昵称
        in: body
        name: NickName
        required: true
        schema:
          type: string
      - description: 头像地址
        in: body
        name: Avatar
        schema:
          type: string
      - description: 部门id
        in: body
        name: DeptId
        schema:
          type: string
      - description: 邮箱
        in: body
        name: Email
        schema:
          type: string
      - description: 所属状态是否有效  1是有效 0是失效
        in: body
        name: Status
        schema:
          type: integer
      - description: 用户手机号码
        in: body
        name: PhoneNum
        schema:
          type: string
      responses:
        "200":
          description: ok
          schema:
            type: string
      tags:
      - 用户
    put:
      description: 修改用户
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: 用户登录账号
        in: body
        name: Account
        required: true
        schema:
          type: string
      - description: 密码
        in: body
        name: PassWord
        required: true
        schema:
          type: string
      - description: 用户显示昵称
        in: body
        name: NickName
        required: true
        schema:
          type: string
      - description: 头像地址
        in: body
        name: Avatar
        schema:
          type: string
      - description: 部门id
        in: body
        name: DeptId
        schema:
          type: string
      - description: 邮箱
        in: body
        name: Email
        schema:
          type: string
      - description: 所属状态是否有效  1是有效 0是失效
        in: body
        name: Status
        schema:
          type: integer
      - description: 用户手机号码
        in: body
        name: PhoneNum
        schema:
          type: string
      responses:
        "200":
          description: ok
          schema:
            type: string
      tags:
      - 用户
  /api/v1/user/{id}:
    get:
      description: 获取单条用户信息
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: ok
          schema:
            type: string
      tags:
      - 用户
swagger: "2.0"